requirements.txt
# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# Data processing
pandas>=1.5.0
numpy>=1.24.0
pillow>=9.5.0
scikit-learn>=1.2.0

# Evaluation metrics
rouge_score>=0.1.2
nltk>=3.8.0
sentencepiece>=0.1.99

# Utilities
tqdm>=4.65.0
protobuf>=3.20.0

# Visualization (optional for Kaggle)
matplotlib>=3.7.0
seaborn>=0.12.0

# Kaggle specific
kagglehub>=0.2.0

# Experiment tracking (optional)
wandb>=0.15.0

# For mixed precision training
accelerate>=0.20.0